Double Int error

+++ Matrix/Matrix/Symmetry.cpp
@@ -5,7 +5,7 @@ int is_symmetric(double matrix[30][30], int R, int C)
 {
     //return 1 if the matrix is symmetric
     //return 0 if the matrix is not symmetric
-    int trans[30][30];
+    double trans[30][30];
     void transpose(double matrix[30][30], int ROW, int COL, double trans[30][30]);
     int is_thesame(double matrix1[30][30], double matrix2[30][30], int R1, int C1, int R2, int C2);
 
@@ -25,8 +25,8 @@ int is_skew_symmetric(double matrix[30][30], int R, int C)
 {
     //return 1 if the matrix is skew-symmetric
     //return 0 if the matrix is not skew-symmetric
-    int trans[30][30];
-    int ans[30][30];
+    double trans[30][30];
+    double ans[30][30];
     void transpose(double matrix[30][30], int ROW, int COL, double trans[30][30]);
     int is_thesame(double matrix1[30][30], double matrix2[30][30], int R1, int C1, int R2, int C2);
     int scalar_multiply(double K, double matrix[30][30], int R, int C, double ans[30][30]);
